Set video scale type to float

+++ android/src/main/java/org/m4m/android/AndroidMediaObjectFactory.java
@@ -160,7 +160,7 @@ public class AndroidMediaObjectFactory implements IAndroidMediaObjectFactory {
     }
 
     @Override
-    public VideoTimeScaler createVideoTimeScaler(int timeScale, FileSegment segment) {
+    public VideoTimeScaler createVideoTimeScaler(float timeScale, FileSegment segment) {
         return new VideoTimeScaler(new MediaCodecEncoderPlugin("video/avc", getEglUtil()), this, timeScale, segment);
     }
